# employee/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Employee API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost/employees/employee/public/",
            "description": "Production"
        }
    ],
    "paths": {
...
        "/api/salaries": {
            "get": {
                "tags": [
                    "Salary"
                ],
                "summary": "Get all salaries",
                "operationId": "App\\Http\\Controllers\\SalaryController::salariesList",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Could Not Find Resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Salary"
                ],
                "summary": "Add a new salary",
                "operationId": "App\\Http\\Controllers\\SalaryController::addSalary",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "employee number": {
                                        "type": "integar"
                                    },
                                    "amount": {
                                        "type": "double"
                                    },
                                    "from_date": {
                                        "type": "date"
                                    },
                                    "to_date": {
                                        "type": "date"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "emp_no": 1,
                                    "amount": 10.5,
                                    "from_date": "2020-10-10",
                                    "to_date": "2020-08-09"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Could Not Find Resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/salaries/{id}": {
            "put": {
                "tags": [
                    "Salary"
                ],
                "summary": "Update salary",
                "operationId": "App\\Http\\Controllers\\SalaryController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Salary id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "employee number": {
                                        "type": "integar"
                                    },
                                    "amount": {
                                        "type": "double"
                                    },
                                    "from_date": {
                                        "type": "date"
                                    },
                                    "to_date": {
                                        "type": "date"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "emp_no": 1,
                                    "amount": 10.5,
                                    "from_date": "2020-10-10",
                                    "to_date": "2020-08-09"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Could Not Find Resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Salary"
                ],
                "summary": "Delete salary",
                "operationId": "App\\Http\\Controllers\\SalaryController::delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Salary id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Could Not Find Resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        }
    }
}